Disable Assist completely
Function calling isn't supported in this branch with an older version of Yandex.Cloud SDK, so the Assist feature must be disabled.

Previously, it made sense to allow Assist to be enabled because, when enabled, the Assist API would append device states to the prompt. However, in newer versions of Home Assistant, the GetLiveContext tool is used to fetch device states instead.